+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Comment model
+- Comment model.
+
+### Changed
+
+- Make python-magic an optional dependency.
 
 
 ## [0.0.31] - 2025-10-23